some updates

+++ Controllers/ItemController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using TodoApi.Middlewares;
 using TodoApi.Models;
 
@@ -10,32 +11,56 @@ namespace TodoApi.Controllers
     public class ItemController : ControllerBase
     {
         private readonly ApiDbContext _context;
+        private readonly ILogger<ItemController> _logger;
 
-        public ItemController(ApiDbContext context)
+        public ItemController(ApiDbContext context, ILogger<ItemController> logger)
         {
             _context = context;
+            _logger = logger;
         }
 
 
         [HttpGet]
         [AuthMiddleware]
-        public ActionResult<IEnumerable<Item>> GetItems()
+        public ActionResult<IEnumerable<Item>> GetItemsFromATodoList([FromQuery] int todolist_id)
         {
-            var user = HttpContext.Items["User"];
-            var items =_context.Item.ToList();
-            return Ok(items);
+            TodoList? todoList = _context.TodoList.Include(tl => tl.Items).FirstOrDefault(tl => tl.Id == todolist_id);
+
+            if (todoList == null)
+            {
+                _logger.LogDebug($"TodoList {todoList?.Id} not found");
+                return Forbid();
+            }
+
+            User? user = HttpContext.Items["User"] as User;
+
+            if (todoList.UserId != user?.Id)
+            {
+                _logger.LogDebug($"Unauthorized access attempt to TodoList {todoList?.Id} by User {user?.Id}");
+                return Forbid();
+            }
+
+            return Ok(todoList.Items);
         }
 
         [HttpGet("{id}")]
         [AuthMiddleware]
         public ActionResult<Item> GetItem(int id)
         {
-            var item = _context.Item.Find(id);
+            Item? item = _context.Item.Include(i => i.TodoListId).FirstOrDefault(i => i.Id == id);
 
             if (item == null)
             {
-                Console.WriteLine("GetItemById NOT FOUND");
-                return NotFound();
+                _logger.LogDebug($"Item {item?.Id} not found");
+                return Forbid();
+            }
+
+            User? user = HttpContext.Items["User"] as User;
+
+            if (item.TodoList?.UserId != user?.Id)
+            {
+                _logger.LogDebug($"Unauthorized access attempt to Item {item?.Id} by User {user?.Id}");
+                return Forbid();
             }
 
             return Ok(item);
@@ -43,39 +68,81 @@ namespace TodoApi.Controllers
 
         [HttpPost]
         [AuthMiddleware]
-        public ActionResult<Item> PostItem(Item item)
+        public ActionResult<Item> PostItem([FromBody] Item item)
         {
-            var newItem = _context.Item.Add(item);
-            _context.SaveChanges();          
-            return Ok(newItem);
+            TodoList? todoList = _context.TodoList.FirstOrDefault(tl => tl.Id == item.TodoListId);
+
+            if (todoList == null)
+            {
+                _logger.LogDebug($"Todo list {todoList?.Id} not found");
+                return Forbid();
+            }
+
+            User? user = HttpContext.Items["User"] as User;
+
+            if (todoList.UserId != user?.Id)
+            {
+                _logger.LogDebug($"Unauthorized attempt to add item to TodoList {todoList?.Id} by User {user?.Id}");
+                return Forbid();
+            }
+
+            _context.Item.Add(item);
+            _context.SaveChanges();
+
+            return Ok(item);
         }
 
         [HttpPut("{id}")]
         [AuthMiddleware]
-        public ActionResult<Item>? PutItem(int id, Item modifiedItem)
+        public ActionResult<Item> PutItem(int id, [FromBody] Item modifiedItem)
         {
-            var item = _context.Item.Find(id);
-            
+            Item? item = _context.Item.Include(i => i.TodoList).FirstOrDefault(i => i.Id == id);
+
             if (item == null)
-                return null;
-            
+            {
+                _logger.LogDebug($"Item {item?.Id} not found");
+                return Forbid();               
+            }
+
+            User? user = HttpContext.Items["User"] as User;
+
+            if (item.TodoList?.UserId != user?.Id)
+            {
+                _logger.LogDebug($"Unauthorized attempt to update item {item.Id} by User {user?.Id}");
+                return Forbid();                
+            }
+
             item.Content = modifiedItem.Content;
-            item.IsComplete = modifiedItem.IsComplete;          
+            item.IsComplete = modifiedItem.IsComplete;
             _context.SaveChanges();
+
             return Ok(item);
         }
 
         [HttpDelete("{id}")]
         [AuthMiddleware]
-        public void DeleteItem(int id)
+        public IActionResult DeleteItem(int id)
         {
-            var item = _context.Item.Find(id);
+            var item = _context.Item.Include(i => i.TodoList).FirstOrDefault(i => i.Id == id);
 
             if (item == null)
-                throw new Exception("Item doesn't exist");
+            {
+                _logger.LogDebug($"Item {item?.Id} not found");
+                return Forbid();               
+            }
+
+            User? user = HttpContext.Items["User"] as User;
+
+            if (item.TodoList?.UserId != user?.Id)
+            {
+                _logger.LogDebug($"Unauthorized attempt to update item {item.Id} by User {user?.Id}");
+                return Forbid();                
+            }
 
             _context.Item.Remove(item);
             _context.SaveChanges();
+
+            return NoContent();
         }
     }
 }
